feat: move AnalyticsIdentifierResult to new data module and update its usages
